done with form and alert

+++ logixtek-intern-sept/src/app/applicantMain/applicant-list/applicant-list.component.ts
@@ -1,11 +1,13 @@
-import { Component, OnInit, Output, EventEmitter, Input } from "@angular/core";
-import { Applicant } from "../../applicant_data/applicant.model";
-import { ApplicantService } from "../service/applicant.service";
+import { Component, OnInit, Output, EventEmitter, Input } from '@angular/core';
+import { Applicant } from '../../applicant_data/applicant.model';
+import { ApplicantService } from '../service/applicant.service';
+import { ConfirmationDialogService } from '../../utils/confirmation-dialog/confirmation-dialog.service';
+
 @Component({
-  selector: "app-applicant-list",
-  templateUrl: "./applicant-list.component.html",
-  styleUrls: ["./applicant-list.component.css"],
-  providers: [ApplicantService]
+  selector: 'app-applicant-list',
+  templateUrl: './applicant-list.component.html',
+  styleUrls: ['./applicant-list.component.css'],
+  providers: [ApplicantService, ConfirmationDialogService]
 })
 export class ApplicantListComponent implements OnInit {
   @Output()
@@ -16,6 +18,9 @@ export class ApplicantListComponent implements OnInit {
 
   @Input()
   newApplicantToList: Applicant;
+
+  @Input()
+  alertStatus: string;
 
   // applicantTest = new Applicant(
   //   'Test 1',
@@ -29,12 +34,15 @@ export class ApplicantListComponent implements OnInit {
   applicants: Applicant[] = [];
   nextId = 0;
 
-  constructor(private applicantService: ApplicantService) {}
+  constructor(
+    private applicantService: ApplicantService,
+    private confirmationDialogService: ConfirmationDialogService
+  ) {}
 
   ngOnInit() {
     // this.applicants = this.applicantService.getApplicant();
     this.applicants = this.applicantService.getApplicants();
-    console.log(this.applicants);
+    console.log(this.alertStatus);
   }
 
   // ================================================
@@ -61,14 +69,29 @@ export class ApplicantListComponent implements OnInit {
   }
 
   deleteApplicant(index: number) {
-    this.applicantService.deleteApplicant(index);
-    // Reload the list
-    this.applicants = this.applicantService.getApplicants();
+    this.confirmationDialogService
+      .confirm(
+        'Delete Confirmation',
+        'Do you really want to delete Applicant with ID = ',
+        this.applicants[index - 1].id + ' '
+      )
+      .then(confirmed => {
+        if (confirmed) {
+          this.applicantService.deleteApplicant(index);
+          // Reload the list
+          this.applicants = this.applicantService.getApplicants();
+          this.alertStatus = 'delete';
+        }
+      })
+      .catch(() =>
+        console.log(
+          'User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'
+        )
+      );
   }
 
   onUpdateApplicant(index: number) {
     const tempApplicant = this.applicantService.selectApplicant(index);
-    tempApplicant.id = index;
     this.applicantSelected.emit(tempApplicant);
   }
 }
